K-means: Fix k to be of type integer and add warnings for when k < number of supplied centroids
JIRA: MADLIB-697, MADLIB-712
Fixed kmeans seeding functions and closest_columns to allow number
of centroids k to be of type integer.  Added parameter checks to
ensure k is within a valid range.  Added warnings to seeding
functions for when k < number of supplied centroids.

+++ src/modules/linalg/metric.cpp
@@ -208,7 +208,7 @@ AnyType
 closest_columns::run(AnyType& args) {
     MappedMatrix M = args[0].getAs<MappedMatrix>();
     MappedColumnVector x = args[1].getAs<MappedColumnVector>();
-    uint16_t num = args[2].getAs<uint16_t>();
+    uint32_t num = args[2].getAs<uint32_t>();
     FunctionHandle dist = args[3].getAs<FunctionHandle>()
         .unsetFunctionCallOptions(FunctionHandle::GarbageCollectionAfterCall);
